Consolidate and remove dead code
