+++ .travis.yml
@@ -63,18 +63,18 @@ script:
 # the coverage collection.
 - go test ./...
 # Also run benchmarks
-- go test -bench . ./search/...
+- go test -bench . ./geolite2/...
 
 # Run all tests, collecting coverage data.
+- go test -covermode=count -coverprofile=common.cov -v github.com/m-lab/annotation-service/common
 - go test -covermode=count -coverprofile=geoip.cov -v github.com/m-lab/annotation-service/geoip
 - go test -covermode=count -coverprofile=handler.cov -v github.com/m-lab/annotation-service/handler
 - go test -covermode=count -coverprofile=loader.cov -v github.com/m-lab/annotation-service/loader
 - go test -covermode=count -coverprofile=metrics.cov -v github.com/m-lab/annotation-service/metrics
-- go test -covermode=count -coverprofile=parser.cov -v github.com/m-lab/annotation-service/parser
-- go test -covermode=count -coverprofile=search.cov -v github.com/m-lab/annotation-service/search
+- go test -covermode=count -coverprofile=geolite2.cov -v github.com/m-lab/annotation-service/geolite2
 
 # Coveralls
-- $HOME/gopath/bin/gocovmerge handler.cov geoip.cov loader.cov parser.cov search.cov > merge.cov
+- $HOME/gopath/bin/gocovmerge handler.cov geoip.cov loader.cov geolite2.cov common.cov > merge.cov
 - $HOME/gopath/bin/goveralls -coverprofile=merge.cov -service=travis-ci
 
 # Build and prepare for deployment
