+++ geolite2/testdata/GeoLite2CityCORRUPT.zip
