+++ geolite2/testdata/GeoLite2CityCORRUPT/GeoLite2-City-Blocks-IPv4.csv
