+++ geolite2/testdata/GeoLite2CityCORRUPT/GeoLite2-City-Locations-en.csv
