+++ geolite2/whitebox_test.go
@@ -1,4 +1,4 @@
-package parser
+package geolite2
 
 // Export just for testing.
 var ConvertIPNodeToGeoData = convertIPNodeToGeoData
