+++ handler/pubsub.go
@@ -6,6 +6,7 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/m-lab/annotation-service/geolite2"
 	"golang.org/x/net/context"
 
 	"cloud.google.com/go/pubsub"
@@ -35,8 +36,7 @@ func waitForDownloaderMessages() {
 	}
 	// Block forever to listen for new messages and run the refresh dataset callbacks when a new message arrives
 	log.Fatal(sub.Receive(context.Background(), func(ctx context.Context, m *pubsub.Message) {
-		PopulateLatestData()
+		geolite2.PopulateLatestData()
 		m.Ack()
 	}))
-
 }
